refactor: Long amount 대신 LottoAmount를 사용하도록 변경

+++ src/main/kotlin/lotto/domain/LottoProfitRateCalculator.kt
@@ -7,6 +7,6 @@ object LottoProfitRateCalculator {
     ): Double {
         val purchaseAmount = userLottos.calculatePurchaseAmount()
         val prizeAmount = lottoResults.calculateTotalPrizeAmount()
-        return (prizeAmount / purchaseAmount.toDouble())
+        return prizeAmount / purchaseAmount
     }
 }
